daemon/requirements.txt
pandas==2.1.1
SQLAlchemy
openpyxl
shapely==1.5.8
psycopg2-binary
xlrd>=2.0.1